Extract out new module `AssetId`.
