+++ lib/coin-selection/lib/Cardano/CoinSelection.hs
@@ -75,6 +75,9 @@ import Cardano.CoinSelection.Size
 import Cardano.CoinSelection.UTxOSelection
     ( UTxOSelection
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -82,8 +85,7 @@ import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle
     )
 import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId
-    , TokenMap
+    ( TokenMap
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
     ( TokenQuantity
