+++ lib/coin-selection/lib/Cardano/CoinSelection/Balance.hs
@@ -139,6 +139,9 @@ import Cardano.CoinSelection.UTxOSelection
 import Cardano.Numeric.Util
     ( padCoalesce
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -146,8 +149,7 @@ import Cardano.Wallet.Primitive.Types.TokenBundle
     ( TokenBundle (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId
-    , Lexicographic (..)
+    ( Lexicographic (..)
     , TokenMap
     )
 import Cardano.Wallet.Primitive.Types.TokenQuantity
