+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/AssetId.hs
@@ -1,0 +1,34 @@
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingStrategies #-}
+
+module Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    ) where
+
+import Prelude
+
+import Cardano.Wallet.Primitive.Types.TokenName
+    ( TokenName
+    )
+import Cardano.Wallet.Primitive.Types.TokenPolicyId
+    ( TokenPolicyId
+    )
+import Control.DeepSeq
+    ( NFData
+    )
+import GHC.Generics
+    ( Generic
+    )
+
+-- | A combination of a token policy identifier and a token name that can be
+--   used as a compound identifier.
+--
+data AssetId = AssetId
+    { tokenPolicyId
+        :: !TokenPolicyId
+    , tokenName
+        :: !TokenName
+    }
+    deriving stock (Eq, Generic, Ord, Read, Show)
+
+instance NFData AssetId
