+++ lib/primitive/lib/Cardano/Wallet/Primitive/Types/TokenBundle.hs
@@ -17,9 +17,8 @@
 --
 module Cardano.Wallet.Primitive.Types.TokenBundle
     (
-    -- * Types
+    -- * Type
       TokenBundle (..)
-    , AssetId (..)
 
     -- * Construction
     , empty
@@ -79,12 +78,14 @@ import Prelude hiding
 import Algebra.PartialOrd
     ( PartialOrd (..)
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
-    , Flat (..)
+    ( Flat (..)
     , Lexicographic (..)
     , Nested (..)
     , TokenMap
