+++ lib/primitive/test/spec/Cardano/Wallet/Primitive/Types/TxSpec.hs
@@ -17,11 +17,11 @@ module Cardano.Wallet.Primitive.Types.TxSpec
 
 import Prelude
 
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenMap.Gen
     ( genAssetId
