+++ lib/wallet/integration/src/Test/Integration/Scenario/API/Shelley/CoinSelections.hs
@@ -38,11 +38,11 @@ import Cardano.Wallet.Api.Types
 import Cardano.Wallet.Primitive.NetworkId
     ( HasSNetworkId
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName (..)
