+++ lib/wallet/mock-token-metadata/src/Cardano/Wallet/TokenMetadata/MockServer.hs
@@ -31,11 +31,11 @@ import Prelude
 import Cardano.Wallet.Primitive.Types
     ( TokenMetadataServer (..)
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenMetadata
     ( AssetDecimals (..)
