+++ lib/wallet/src/Cardano/Wallet/Address/Keys/MintBurn.hs
@@ -44,11 +44,11 @@ import Cardano.Wallet.Flavor
 import Cardano.Wallet.Primitive.Passphrase
     ( Passphrase (..)
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName
