+++ lib/wallet/src/Cardano/Wallet/Primitive/Types/Tx/TxSeq.hs
@@ -113,6 +113,9 @@ import Prelude hiding
 import Cardano.Wallet.Primitive.Model
     ( applyTxToUTxO
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -121,9 +124,6 @@ import Cardano.Wallet.Primitive.Types.Hash
     )
 import Cardano.Wallet.Primitive.Types.StateDeltaSeq
     ( StateDeltaSeq
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenName
     ( TokenName (..)
