+++ lib/wallet/test/unit/Cardano/Wallet/Shelley/TransactionSpec.hs
@@ -104,6 +104,9 @@ import Cardano.Wallet.Primitive.Passphrase
 import Cardano.Wallet.Primitive.Types.Address
     ( Address (..)
     )
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -119,8 +122,7 @@ import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenBundle
-    ( AssetId
-    , TokenBundle
+    ( TokenBundle
     )
 import Cardano.Wallet.Primitive.Types.TokenBundle.Gen
     ( genTokenBundleSmallRange
@@ -1618,7 +1620,7 @@ instance Arbitrary AnyRecentEra where
 
 instance Arbitrary AssetId where
     arbitrary =
-        TokenBundle.AssetId
+        AssetId
         <$> arbitrary
         -- In the calculation of the size of the Tx, the minting of assets
         -- increases the size of the Tx by both a constant factor per asset
