+++ lib/wallet/test/unit/Cardano/Wallet/TokenMetadataSpec.hs
@@ -9,11 +9,11 @@ module Cardano.Wallet.TokenMetadataSpec
 
 import Prelude
 
+import Cardano.Wallet.Primitive.Types.AssetId
+    ( AssetId (..)
+    )
 import Cardano.Wallet.Primitive.Types.Hash
     ( Hash (..)
-    )
-import Cardano.Wallet.Primitive.Types.TokenMap
-    ( AssetId (..)
     )
 import Cardano.Wallet.Primitive.Types.TokenMetadata
     ( AssetDecimals (..)
